Move relay list logic to RelayListContext and scroll logic to ScrollPositionContext
